+++ StereoPlus2/Math.hpp
@@ -181,32 +181,30 @@ struct Convert {
 	// Millimeters to [-1;1]
 	// World center-centered
 	// (0;0;0) in view coordinates corresponds to (0;0;0) in world coordinates
-	static glm::vec3 MillimetersToViewCoordinates(const glm::vec3& vMillimeters, const glm::vec2& viewSizePixels, const float& viewSizeZMillimeters) {
-		static float inchToMillimeter = 0.0393701;
-		auto vsph = viewSizePixels / 2.f;
-		auto vszmh = viewSizeZMillimeters / 2.f;
-
-		auto vView = glm::vec3(
-			vMillimeters.x * Settings::PPI().Get() * inchToMillimeter / vsph.x,
-			vMillimeters.y * Settings::PPI().Get() * inchToMillimeter / vsph.y,
-			vMillimeters.z / vszmh
-		);
-		return vView;
-	}
-	// Millimeters to [-1;1]
-	// World center-centered
-	// (0;0;0) in view coordinates corresponds to (0;0;0) in world coordinates
-	static glm::vec3 MillimetersToViewCoordinates(const glm::vec3& vMillimeters, const glm::vec2& viewSizePixels) {
+	static glm::vec2 MillimetersToViewCoordinates(const glm::vec3& vMillimeters, const glm::vec2& viewSizePixels) {
 		static float inchToMillimeter = 0.0393701;
 		auto vsph = viewSizePixels / 2.f;
 		
-		auto vView = glm::vec3(
+		return glm::vec2(
+			vMillimeters.x * Settings::PPI().Get() * inchToMillimeter / vsph.x,
+			vMillimeters.y * Settings::PPI().Get() * inchToMillimeter / vsph.y
+		);
+	}
+
+	// Millimeters to [-1;1]
+	// World center-centered
+	// (0;0;0) in view coordinates corresponds to (0;0;0) in world coordinates
+	static glm::vec3 MillimetersToViewCoordinates3(const glm::vec3& vMillimeters, const glm::vec2& viewSizePixels) {
+		static float inchToMillimeter = 0.0393701;
+		auto vsph = viewSizePixels / 2.f;
+
+		return glm::vec3(
 			vMillimeters.x * Settings::PPI().Get() * inchToMillimeter / vsph.x,
 			vMillimeters.y * Settings::PPI().Get() * inchToMillimeter / vsph.y,
-			0
+			vMillimeters.z
 		);
-		return vView;
 	}
+
 	// Millimeters to [-1;1]
 	// World center-centered
 	// (0;0;0) in view coordinates corresponds to (0;0;0) in world coordinates
@@ -249,32 +247,20 @@ struct Convert {
 };
 
 class Stereo {
-	static glm::vec3 getLeft(const glm::vec3& posMillimeters, const glm::vec3& cameraPos, float eyeToCenterDistance, const glm::vec2& viewSize, float viewSizeZ) {
-		auto pos = Convert::MillimetersToViewCoordinates(posMillimeters, viewSize, viewSizeZ);
-		float denominator = cameraPos.z - pos.z;
-		return glm::vec3(
-			(pos.x * cameraPos.z - pos.z * (cameraPos.x - eyeToCenterDistance)) / denominator,
-			(cameraPos.z * -pos.y + cameraPos.y * pos.z) / denominator,
-			0
-		);
-	}
-	static glm::vec3 getRight(const glm::vec3& posMillimeters, const glm::vec3& cameraPos, float eyeToCenterDistance, const glm::vec2& viewSize, float viewSizeZ) {
-		auto pos = Convert::MillimetersToViewCoordinates(posMillimeters, viewSize, viewSizeZ);
-		float denominator = cameraPos.z - pos.z;
-		return glm::vec3(
-			(pos.x * cameraPos.z - pos.z * (cameraPos.x + eyeToCenterDistance)) / denominator,
-			(cameraPos.z * -pos.y + cameraPos.y * pos.z) / denominator,
-			0
-		);
+	static const glm::vec3& getStereoPair(const glm::vec3& posMillimeters, const glm::vec3& cameraPos, float eyeToCenterDistance, const glm::vec2& viewSize) {
+		auto pos = Convert::MillimetersToViewCoordinates(posMillimeters, viewSize);
+		float denominator = cameraPos.z - posMillimeters.z;
+
+		auto xl = (pos.x * cameraPos.z - posMillimeters.z * (cameraPos.x - eyeToCenterDistance)) / denominator;
+		auto xr = (pos.x * cameraPos.z - posMillimeters.z * (cameraPos.x + eyeToCenterDistance)) / denominator;
+		auto y = (cameraPos.z * -pos.y + cameraPos.y * posMillimeters.z) / denominator;
+
+		return glm::vec3(xl,xr,y);
 	}
 public:
-	static glm::vec3 GetLeft(const glm::vec3& v, const glm::vec3& cameraPos, float eyeToCenterDistance, const glm::vec2& viewSize, float viewSizeZ) {
-		return getLeft(v, cameraPos, eyeToCenterDistance, viewSize, viewSizeZ);
+	static const glm::vec3& GetStereoPair(const glm::vec3& v, const glm::vec3& cameraPos, float eyeToCenterDistance, const glm::vec2& viewSize) {
+		return getStereoPair(v, cameraPos, eyeToCenterDistance, viewSize);
 	}
-	static glm::vec3 GetRight(const glm::vec3& v, const glm::vec3& cameraPos, float eyeToCenterDistance, const glm::vec2& viewSize, float viewSizeZ) {
-		return getRight(v, cameraPos, eyeToCenterDistance, viewSize, viewSizeZ);
-	}
-
 };
 
 class Build {